requirements.txt
python_pptx==0.6.19
Wand==0.6.6
